backend - added sign up and sign in.
