+++ backend/backend-notes.txt
@@ -1,0 +1,3 @@
+changes to package.json
+- added "type": "module" (ES6)
+- run: npm run dev instead of nodemon
