+++ backend/package.json
@@ -3,9 +3,10 @@
     "version": "1.0.0",
     "description": "Express.js backend API for ReMap memory sharing application",
     "main": "src/server.js",
+    "type": "module",
     "scripts": {
         "start": "node src/server.js",
-        "dev": "nodemon src/server.js",
+        "dev": "nodemon --env-file=.env src/server.js",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "keywords": [
@@ -18,15 +19,16 @@
     "author": "ReMap Development Team",
     "license": "ISC",
     "dependencies": {
+        "@supabase/supabase-js": "^2.49.8",
         "cors": "^2.8.5",
-        "dotenv": "^16.3.1",
+        "dotenv": "^16.5.0",
         "express": "^4.21.2",
         "helmet": "^7.1.0",
         "morgan": "^1.10.0",
         "pg": "^8.11.3"
     },
     "devDependencies": {
-        "nodemon": "^3.0.2"
+        "nodemon": "^3.1.10"
     },
     "engines": {
         "node": ">=20.0.0",
