+++ backend/src/config/supabaseClient.js
@@ -1,0 +1,7 @@
+
+import { createClient } from '@supabase/supabase-js';
+// const supabaseUrl = 
+// const supabaseKey = 
+const supabase = createClient(supabaseUrl, supabaseKey);
+
+export default supabase;
