+++ frontend/package.json
@@ -15,12 +15,13 @@
     "preset": "jest-expo"
   },
   "dependencies": {
+    "@expo/ngrok": "^4.1.3",
     "@expo/vector-icons": "^14.1.0",
     "@react-native-async-storage/async-storage": "2.1.2",
     "@react-navigation/native": "^7.1.6",
     "@react-three/fiber": "^9.1.2",
-    "@supabase/supabase-js": "^2.49.8",
-    "expo": "~53.0.9",
+    "@supabase/supabase-js": "2.49.5-next.1",
+    "expo": "~53.0.0",
     "expo-constants": "^17.1.6",
     "expo-font": "~13.3.1",
     "expo-gl": "~15.1.5",
